@@ -13,72 +13,72 @@
     "require": {
         "php": ">=8.1.0",
         "academe/omnipay-authorizenetapi": "3.1.2",
-        "adodb/adodb-php": "5.22.6",
-        "aranyasen/hl7": "3.1.1",
-        "bacon/bacon-qr-code": "2.0.7",
-        "digitickets/lalit": "3.3.0",
-        "dompdf/dompdf": "2.0.3",
-        "ezyang/htmlpurifier": "4.16.0",
-        "google/apiclient": "2.15.1",
-        "guzzlehttp/guzzle": "7.8.0",
-        "html2text/html2text": "4.3.1",
-        "kamermans/guzzle-oauth2-subscriber": "1.0.13",
-        "knplabs/knp-snappy": "1.4.4",
+        "adodb/adodb-php": "5.22.7",
+        "aranyasen/hl7": "3.1.5",
+        "bacon/bacon-qr-code": "3.0.1",
+        "digitickets/lalit": "3.4.0",
+        "dompdf/dompdf": "3.0.0",
+        "ezyang/htmlpurifier": "4.17.0",
+        "google/apiclient": "2.18.0",
+        "guzzlehttp/guzzle": "7.9.2",
+        "html2text/html2text": "4.3.2",
+        "kamermans/guzzle-oauth2-subscriber": "1.1.0",
+        "knplabs/knp-snappy": "1.5.0",
         "laminas/laminas-config": "3.9.0",
-        "laminas/laminas-db": "2.18.0",
-        "laminas/laminas-eventmanager": "3.12.0",
+        "laminas/laminas-db": "2.20.0",
+        "laminas/laminas-eventmanager": "3.13.1",
         "laminas/laminas-escaper": "2.13.0",
-        "laminas/laminas-filter": "2.33.0",
-        "laminas/laminas-form": "3.13.1",
-        "laminas/laminas-inputfilter": "2.28.0",
+        "laminas/laminas-filter": "2.38.0",
+        "laminas/laminas-form": "3.21.0",
+        "laminas/laminas-inputfilter": "2.30.1",
         "laminas/laminas-json": "3.6.0",
-        "laminas/laminas-json-server": "3.8.0",
-        "laminas/laminas-loader": "2.10.0",
-        "laminas/laminas-mvc": "3.6.1",
-        "laminas/laminas-mvc-i18n": "1.7.0",
+        "laminas/laminas-json-server": "3.9.0",
+        "laminas/laminas-loader": "2.11.0",
+        "laminas/laminas-mvc": "3.7.0",
+        "laminas/laminas-mvc-i18n": "1.9.0",
         "laminas/laminas-servicemanager": "3.22.1",
         "laminas/laminas-soap": "2.13.0",
-        "laminas/laminas-stdlib": "3.18.0",
-        "laminas/laminas-xmlrpc": "2.17.0",
+        "laminas/laminas-stdlib": "3.19.0",
+        "laminas/laminas-xmlrpc": "2.20.0",
         "lcobucci/jwt": "4.2.1",
         "league/csv": "9.8.0",
         "league/omnipay": "3.2.1",
-        "mobiledetect/mobiledetectlib": "3.74.3",
-        "monolog/monolog": "3.5.0",
-        "mpdf/mpdf": "8.2.0",
-        "nyholm/psr7": "1.8.0",
-        "nyholm/psr7-server": "1.0.2",
+        "mobiledetect/mobiledetectlib": "4.8.06",
+        "monolog/monolog": "3.7.0",
+        "mpdf/mpdf": "8.2.4",
+        "nyholm/psr7": "1.8.2",
+        "nyholm/psr7-server": "1.1.0",
         "omnipay/stripe": "3.2.0",
         "openemr/mustache": "2.15.2",
         "particle/validator": "2.3.5",
-        "pear/archive_tar": "1.4.14",
-        "phpmailer/phpmailer": "6.8.1",
-        "phpoffice/phpspreadsheet": "1.29.0",
-        "phpseclib/phpseclib": "3.0.36",
-        "php81_bc/strftime": "0.5.0",
-        "php-http/discovery": "1.19.1",
-        "psr/log": "2.0.0",
-        "ramsey/uuid": "4.7.4",
+        "pear/archive_tar": "1.5.0",
+        "phpmailer/phpmailer": "6.9.2",
+        "phpoffice/phpspreadsheet": "3.3.0",
+        "phpseclib/phpseclib": "3.0.42",
+        "php81_bc/strftime": "0.7.5",
+        "php-http/discovery": "1.20.0",
+        "psr/log": "3.0.2",
+        "ramsey/uuid": "4.7.6",
         "robthree/twofactorauth": "2.0.0",
         "rospdf/pdf-php": "0.12.67",
-        "smarty/smarty": "4.3.4",
+        "smarty/smarty": "4.5.4",
         "steverhoades/oauth2-openid-connect-server": "2.6.1",
-        "stripe/stripe-php": "13.2.0",
-        "symfony/config": "5.4.26",
-        "symfony/console": "5.4.28",
-        "symfony/dependency-injection": "5.4.29",
-        "symfony/event-dispatcher": "5.4.26",
-        "symfony/finder": "5.4.27",
-        "symfony/filesystem": "5.4.25",
-        "symfony/http-client": "5.4.29",
-        "symfony/http-foundation": "5.4.30",
-        "symfony/yaml": "5.4.30",
-        "twig/twig": "3.7.1",
-        "vlucas/phpdotenv": "5.5.0",
+        "stripe/stripe-php": "15.10.0",
+        "symfony/config": "6.4.8",
+        "symfony/console": "6.4.12",
+        "symfony/dependency-injection": "6.4.12",
+        "symfony/event-dispatcher": "6.4.8",
+        "symfony/finder": "6.4.11",
+        "symfony/filesystem": "6.4.12",
+        "symfony/http-client": "6.4.12",
+        "symfony/http-foundation": "6.4.12",
+        "symfony/yaml": "6.4.12",
+        "twig/twig": "3.14.0",
+        "vlucas/phpdotenv": "5.6.1",
         "waryway/php-traits-library": "1.0.4",
         "yubico/u2flib-server": "1.0.2",
-        "twilio/sdk": "7.16.2",
-        "ringcentral/ringcentral-php": "3.0.1"
+        "twilio/sdk": "8.3.4",
+        "ringcentral/ringcentral-php": "3.0.3"
     },
     "config": {
         "platform": {
