@@ -22,7 +22,7 @@ require_once "$srcdir/clinical_rules.php";
 require_once "$srcdir/report_database.inc.php";
 
 use OpenEMR\Common\Acl\AclMain;
-use OpenEMR\ClinicialDecisionRules\AMC\CertificationReportTypes;
+use OpenEMR\ClinicalDecisionRules\AMC\CertificationReportTypes;
 use OpenEMR\Common\Csrf\CsrfUtils;
 use OpenEMR\Common\Twig\TwigContainer;
 use OpenEMR\Services\PractitionerService;
